Restructure the date input for events
Events could not select the date and time (duration).

This commit is an intermediate commit to add date and time
fields to the event dialog.  Times can now be picked from
a list. Date pickers are next.

The previous implementation used HTML5 datepicker (input type=date)
which is not supported in many browsers and was a hack.