Functional date picking UI
Can now pick dates and events, and dates automatically
adjust based on what's selected.  Uses the
bootstrap-datetimepicker library, which relies on
momentjs (which looks incredibly useful for time
manipulation in the future).

Now just need to hook the dates up to the event model
changes, so the dates get stored correctly on the backend.