Update madlibs; fix new task responsiveness
This commit addresses the updates requested in Issue #56.
The workflow is now the same in the new opportunity modal,
the task show view, and the task edit view.

I've moved the description to the end as per issue #56,
however it looks weird because the first input has
a lot of white space.

The modal has been updated to be responsive, and each
of the elements automatically resizes to the correct
proportions.

This also addresses issue #49.  At this point, every
component of the application is now responsive and
viewable on any sized device.