Commit fonts to the repo (woff format)
The fonts are now in the repo and distribution, so that
all of the files can be served from our webserver and do
not require outside internet access.

The font names are in lowercase to address an IE bug
that doesn't work with font url uppercase.

The fonts are only available in WOFF format, which is
supported in IE9+, Firefox, Chrome, and Safari. The
only browser that doesn't support WOFF is Android <= 4.3.
Android 4.4 does support WOFF.  Even Blackberry supports
WOFF.

This addresses issue #45.