Send filename with attachment; store mime type
This allows attachments to be downloaded with the same
name and type as they were uploaded with.
Fixes issue #93.

Also remove the ability to download by file name.  This
would require much more security and we don't use this
anywhere, so get rid of it to prevent problems down the road.