Seperate local & register apis; look up by email
The /api/auth/local api is now used solely for authentication
and will not create an account if it does not exist

The /api/auth/register api will create a new account
if it doesn't exist.  If an account does exist, it will
error out of the process indicating that the account exists.

Logging in and registering now both looks for multiple
email addresses.  This way you can log in with any of the email
addresses registered with your account.

Addresses pull request / issue #105.