Fix unit test regressions
In moving from Sails 0.9.4 to Sails 0.9.8, the response code
for calling POST to create a model in a controller changed
from 200 to 201.  This broke a lot of the unit tests.

The login procedure also changed with the addition of
'register'.  Update the helper to correctly register if the
user login fails.

The project object changed with the addition of projectOwner
objects.  Fix the like test to use the owner object
within the project object correctly.

This should fix all of the current regressions and thus
all the unit tests should pass.