Make edit profile & project behave the same
When editing a profile or a project, successful save should
take you back to the profile or project rather than keeping you
on the edit view.  Also standardize the styling of the buttons
and the text displayed.  Moved to "Discard Changes" for the cancel
button and "Save [Item]" for the save button.