Migrate from Dropzone to jQuery-File-Upload
Dropzone isn't compatible with IE8/9.  Dropzone also
uses jQuery-File-Upload as its base, so let's use the
base instead which supports IE6+.  I wasn't happy
with Dropzone anyway; it would try to pop up too
many of its own UI elements.

To do so, we need to change how file uploads work.
In particular, due to IE sucking so much, we need to
return an HTML response with JSON encoded in it,
and then use the appropriate logic to decode the result.

FileController also updated to handle multiple
concurrent file uploads, and return an object
that has all of the file entries that have
been added.

With this commit, all of the Dropzone elements are now
broken; a future commit will migrate the rest over to
jQuery-File-Upload.

The element that has been converted for testing is
the project_item_view (basically when you use
Project>Edit).  It has been tested in both
Chrome and IE9.

Add HTML5Shiv and Respond.js in the hopes that these
help rendering problems on old IE versions.

This commit partially addresses #130