SSPI seems to be working
Punched some ugly code through to make it work. Not completely pleased with lack of refactoring I was able to do in the user service file.

configurations for sspi are found in the config/settings/auth.js file.
All site actions now route through MainController, which allows global policies to be put in place. The SSPI policy is one of these.