Unify the login configuration
In assets/js/backbone/config, change login.js to login.json.

Import this file both on the backend and the frontend.  This
allows us to pull in the settings for which login services
to enable (and thus is a bit more secure since we can ignore
any login services that aren't specified in login.json).

Modified AuthController so that instead of writing an
endpoint for each type of passportjs target/strategy,
there's one `oauth` endpoint and one `callback`
endpoint.  The parameter passed to those is the strategy
in use, eg LinkedIn.  That strategy must exist in
login.json, otherwise you'll get a 403 error.