Catch PassportJS errors
When an error occurs in PassportJS, we weren't catching it
by providing a `next()` function.  Specify the `next` function
now to catch those errors and report them both to the
browser and to the sails.log.error logger.

In the future, should encapsulate this in JSON and render
the main view again with the error.