Make ESC key work for modals
Turns out that you have to add `tabindex="-1"` to the modal
div if you want the ESC key to work.  I don't know why this
hack works, but here's the GitHub issue that tracks it
in bootstrap:
https://github.com/twbs/bootstrap/issues/4854

The problem with how we are using modals is that
modal_wizard relied on an existing modal div to exist,
whereas regular modal did not.  This is legacy from an
old developer, and needed to be normalized.  Now, you
pass the container div as the `el` and an `id` for the
modal, just like you do with regular modals.

This ensures that the modal HTML is always generated by
modal or modal_wizard, and eliminates writing that HTML
in multiple places which results in bugs or differences
in code.  Consistency is the key.

This resolves #148.