Scope to this components el
When adding new tasks, some of the tags were getting
repeated.  My hypothesis is that some DOM elements
were not being cleaned up.  As good practice, scope
all of the DOM elements in `task_form_view` to only
operate on their given `el` when the view is created.

This seems to fix the issue on my end.

This is related to
https://github.com/USStateDept/midas-crowdwork/issues/22