Add `disableClose` option to modals
Now, in a modal, you can specify as an option in the message
object the `disableClose` flag which will prevent the user
from closing the modal.

Instead of just passing a string, you pass a message object:
{
  message: 'foo',
  disableClose: true
}

Message can include HTML links; if they render within the
application, use the `link-backbone` class which will
handle the navigation withing backbone.

The task controller has been updated to check if you
have priviledges to edit the task, and if not, will
pop up the login modal.  You can either click to the
task, or login.  The close options are disabled.

This fixes #155