Make @ autocomplete work
This is an initial, WIP commit.

In a comment box, you can now use the `@` symbol to
search for users, projects, or wikipedia entries.

Support for saving and parsing the text is not yet provided.

But the UI component works.

This required a number of changes:
- sources now provide a `link` field so you know how to link to it
- removed profile source because usernames are emails and should
  not be exposed
- overrode the At.js library to use underscore templates instead
- added CSS for At.js (basic, not styled)

Things that need to be done:
- Styling for dropdown results (eg, user photo, wikipedia, etc)
- Saving div/textarea data to the database