Save comment to database
Store the comment text in the data object (the HTML) so
that it is sent on to the database and renders correctly
when it comes back.  Currently stored as HTML; in the future
it should be stored as some sort of markup that can make
sure the names/usernames display their current values.
Same thing for project names.

But for now, this is probably good enough.  Saving and
retreiving and rendering all seems to work.