Organize results in config'd order
When multiple sources are queried, the queries are executed
by async in parallel.  Each query may come back at a different
time.  Instead of putting the results in the final array
immediately, put them in a preliminary array and then reassemble
in the correct order once all the queries are complete.

Ensures that the order of the results matches the order
specified in the sources.js config file.