WIP: customize At.js to do sane things
At.js has some wacky defaults, like putting a space after
each selection that isn't delete-able by the user (hate that!)

It was also trying to sort the results; don't do that, the
server has already sorted them for you.

It was trying to highlight in a wacky way; put a placeholder
in for better highlighting.

TODO:
- replace matching function in highlighter
- add non-breaking space after the inserted <span> element