Clean up sspi credentials
Don't need a million variables, reuse and store in req.sspi
while doing authentication.

As per #137, create a new `rawUser` part of the credentials
that provides direct access to the user specified by the
domain header.  This is better than doing a split or other
hacky technique.  Since the `req` object is passed to the
authentication function, use the req object to retreive the
rawUser.

With password validation turned off, can now directly
pass username/password as user@emailSuffix/domain, so effectively
log in with username and domain as password.  This is why we
needed to remove the login/register/logout functions when
sspi is enabled.