Use common task query function
When searching tasks, the results were throwing an
error on render because they were using a different formula
to create the task result object (#157).

Seperate out the task.find() function from the controller
to the utility service.  Then use the utility service with
both the search and task.find endpoints.  The result is
common object structure for both types of queries.