Allow middle click and CMD/CTRL click
With this change, using the middle button (on windows) to
click or holding command/control keys and using the left
mouse button (mac style) will let the link handling
be done by the browser instead of backbone.  This is useful
for opening new windows or tabs when clicking on things.