Format autocomplete dropdown
Put the right image (either project image or
profile image) on the left, and title on
the right.  For wikipedia, use the wikipedia
globe.

Provide description and title/agency
underneath.

Disabled the highlighter; not sure that it is
necessary but it is there if we want to bring it
back again.