Pass alerts to view to render
In the case that a user's account is disabled, and we're
not using normal login dialogs (eg, SSPI), be sure to pass
the message through.

Rename main/find to main/index to match the view.
Update routes to match this target.

Create a data variable to be passed to the template, so
more dynamic data can be passed.

Add a `signin` flag to login.json to check if you even
want to have a sign in link on your website.

Pass req.alert to the view via the data variable.

Update sspi to pass errors separately from invalid users;
invalid users should still render the page but with the
respective warning.