Add unlock and reset password
Add to the admin panel the ability to unlock a user that
has exceeded their allowable password attempts. Added
an admin function in AdminController.

Add to the admin panel the ability to reset a user's
password.  Uses a new `resetPassword` endpoint on the
UserController (/api/user/resetPassword).  If the
user is not an admin, it will reset their password.
Otherwise, it will use an option id parameter posted
to the API which will identify which user to change
the password.

Added the passwordLockout option to auth.ex.js