Set more logical cache options
For files, remove no-cache directives
Instead, require revalidation every 5 minutes.  This should
be fine for things like profile photos, where it is ok if they're
not up to date within 5 minutes.

For the main view, require revalidation every hour.  This file
will rarely update, and so it is okay if it takes an hour to propogate.
In reality, this could probably be set to a day.  But the file is
so small, it doesn't matter.