Version production assets
Production assets are currently called:
- midas.js
- midas.css / midas.min.css

When various webservers set their caching to be very long periods
of time, new releases of midas are not loaded from the server.

With this change, the production assets are now called:
- midas-{ version }.js
- midas-{ version }.min.css
eg:
- midas-0.0.11.js
- midas-0.0.11.min.css

This way, when a new version is released, the filenames will
be udpated.  When the browser hits the page, it will see that
it doesn't have these files in cache, and will loads them.

This is made possible by e7db69e28a97c9aaf6216fa12a8092c226a56efb
which forces the main view to be reloaded hourly.