Consolidate project reply notifications
There was no need for the same code to be implemented almost
exactly the same in both prepareProjectCommentReplyOwnerEmail
and prepareProjectCommentReplyParentEmail.  Combined those
into a single preparation file and updated the settings to
use this strategy.

Update the HTML/Text templates for project comment replies
to match the content that we want.  Update the subject lines
to be underscore templates in the config.