Use the same comment preflight for tasks
Realized that there's no reason to use 4 different files
when 1 file can do the whole thing.  So consolidated the
two 2 task comment preflights into a single
preflight 'prepareCommentReplyEmail'.  Plus it has a
dramatically shorter, and more awesome, file name.

Updated settings file to specify which template to use.

Updated the preflight to check whether the target is a
task or a project, and perform the appropriate model
lookups.