Consolidate comment templates
Both task and project emails use the same basic format.
Instead of having separate email templates for both,
just use one template and check in the template whether
the email is about a task or project.

Update wording in the subject lines to be a bit more
friendly and logical.