Overhaul task volunteer email
When a user volunteers for a task, the owner of that task
gets an email that tells them a new user has volunteered.

Update the email to use the new template style.

Update the preflight to maintain the metadata object.

Update the configuration to use a dynamic subject (_.template)

Pass in the emailName in the notifications template settings
as per the other options (allows changing those from the
settings as opposed to the code)