Add the first phase of forgot password
The first phase includes the endpoing for the UI that
looks up the user account and then generates a password
reset token for the user.

If the server is in test mode, the token will be returned
by the API.  If the server is not in test mode, only
a vague `success:true` with the submitted email address
will be provided.

The next steps:
- Generate notification email
- Create profile/password page for resetting password
- If given a valid token, gray out the old password field
- Backend token verification function