Drastically simplify notifications configuration
There were a ton of settings in notifications that
really didn't do anything.  They were put in there
in the interest of being used at some point in the
future, but were never implemented.  It made the
notifications settings overly confusing and made
it difficult to parse the machinery.  By removing
these settings, we lose absolutely nothing and
we gain a ton of clarity.