Supress deliveries for initiators
If you initiate a comment or volunteer for your own
project, or basically if any notification caused by you
is destined to you, that notification should be supressed.

This commit adds to the preflights an `initiatorId` and
then checks that Id before allowing delivery.  If they
match, it aborts delivery for that particular notification
to that recipient.

You no longer get emails saying you replied to your own project