Truncate titles in cards; make names clickable
If someone had a really long title or agency, it would make
their info wrap outside of the card.  Truncate at one line
instead.

Force the photo to go on the left and not have the text wrap
around it.

Make the cards a little larger to accomodate people who like
to write a lot.

Make the names in the cards clickable so you can navigate to
their profile.

This fixes #168.