Add user names to img alt tags
For screenreaders/accessibility and people who can't
view images/photos, we need to include an alt tag that
has the user's / owner's / volunteer's name.

This commit refactors the `addOwnerName` function into
a user utility, and then calls it from project and task
metadata functions.

Update the rendering of tasks to show the alt text.

Finishes what's requested for #127.