Set select2 width to fix placeholders
Placeholders weren't rendering correctly when they're
initially hidden.  The select2 people say this is
fixed, but it isn't.

The ugly hack is to set the width of each input
explicitly.  I hate mixing CSS with JS, but in this
case there doesn't appear to be any other way to make
it work.

Fixes #161