Correcting the location of the query parameters.
Previously, the query parameters were added to the root
of the fields object. None of the post processing was reading
those properties. The templateLocals parameter is where the data
is pulled from. Also added a missing id to the 'madlibs' on the
task page that was causing that field to not be populated.

Fixes #251.