Allow adding multiple tags with comma separator.
Currently, to enter multiple tags inside a field, such as the Skills
for a Project or an Opportunity, you have to type the tag, then either
click on it from the dropdown or press the tab key, then repeat.

To make it easier to enter multiple tags quickly, Select2 allows you to
define one or more separators, such as a comma, so that you can type
the first tag, press the comma key, and start typing the second tag.
Select2 docs: http://ivaynberg.github.io/select2/#doc-tokenSeparators

In this commit, I’ve added the comma separator functionality for
Skills and Topics only since tags for Location and Agency should be
able to contain commas.

Fixes #304 in GitHub.