Corrected problem with server hang and 'typo in translation'.
i18n was requesting the translation file from an incorrect
location if the user navigated directly to a item-level url.
i.e. Navigating directly to '/projects/1' would cause
i18n to request from 'projects/locales/en/translation.json'
instead of 'locales/en/translation.json'. The problem would
not appear if the user navigated to 'projects' and clicked
on the link for a project.

Corrected by adding a '/' to the beginning of the
resGetPath url to force jQuery to look from the root
instead of the current path.

Actually fixes #280.