Admin csv export for users and tasks
Add a "download" button on the user and tasks admin pages to save a csv
format of those files. CSV file formatted for easy import into Excel by
handling null strings and formatting dates. Adds dependency on the
json2csv module to produce well-formatted csv (handles escape quoting,
etc.)

This fixes issues #667 and #668.

Some clumsiness on the task csv in how it counts volunteers to summarize
in the "signup" column. I couldn't get a groupby/count to work with the
waterline ORM, so had to resort to clunky sql and exclude that for test
cases, where the backend isn't sql-compliant.