handle ajax errors globally with auth special case
Part of fix for #727. Per discussion with @dhcole, instead of handling
auth errors only in the admin section, this general handler helps
for the whole app. Now whenever we get an auth error, fire up login.

Add handling in the nav to prevent the login modal from opening multiple
times.

Instead of displaying a hardcoded error message, display the error we
get back in the entity to the user. So those messages should be worded a
bit better now: "You must login..." instead of "Access denied."