Add clone to TaskController + corresponding test
Implement basic clone endpoint in TaskController that takes a taskId
and creates a new task, copying over the info from the original task.
Also add a functional test covering the new endpoint. The endpoint
still needs added functionality to clone the tags from the old task to
the cloned task.