Implement Clone Task UI for task owners & admins
Add an option tasks allowing them to be cloned. When the option is
selected, show a modal asking what to use as the title for the clone.
Once the user confirms the clone and it completes sucessfully, take
them to the show page for the new cloned task.

Implements issue #734