Rename task clone -> copy at UI level
Task clone is now task copy at code & UI level. Additionally, all users
who are logged in now are shown the Copy Opportunity option. The option
is no longer available in the task editing state. When a task is copied
the user now is sent to the edit page for the new copy.