Filter Required/Not Required from task tag clouds
In the list of tasks shown on a project page, Required/Not Required
tags are not useful to include in the tag cloud rendered for each task.
These are associated with tags of type task-skills-required, which are
now filtered out of the task tag list before each tag is rendered.

This addresses issue #776