people view: simpler, more robust tooltip posn
Previously I was trying to calculate the position of the tooltip by find
ing the position of the dot and figuring out where it was on the screen.
That proved to be fragile due to offsets and SVG scaling.

This method is much simpler: on entering the dot, position relative to
the mouse pointer position. That means the position can subtly change
depending on whether you enter a dot from the top or bottom, say, but it
looks good regardless.