removed spurrious project router
While working on some router stuff I happened to notice that this file
contained some routes that were redundant with those in
browse/browse_app.js, where all the other routes live. And it wasn't
referenced. Looks like a copy/paste problem, or a vestige of a prior
attemt at refactoring the routers?

After removal the tests still pass and the app functions, so that feels
pretty redundant to me.