Fix issue with profile image update in FF
Firefox seems to be caching the first image and not
retrieving the new image when the  CSS is bumped.
Probably because the URL to the image never changes
(for example /api/user/photo/13).
My solution was to put a timestamp as a URL parameter when
retrieving the new image, (ie. /api/user/photo/13?1439337665153).
I believe this busts FF image cache, and forces FF to
get the newly uploaded image.

https://github.com/18F/midas/issues/800