Redline changes for activity feed
These changes are bunched together to
represent a series of redline changes
from Gail for the site. These all apply
to the new activity feed and have to do
with font size, color and spacing.
Since we don't currently have badges
we're just going to show off all the
tasks that get completed with the help
 of users.