Make tests pass with new prompt
Turns out the tags weren't being set by the test user. This had to do
with the way I was instructing casper to register the user at the top
of the test suite. That has been fixed, so now the agency and location
tags do get set on the test user, enabling the tests to pass as normal.