Tests check for array length not error
Per @ultrasaurus’s comments, the tests were previously testing for the
presence of an error object, instead of just checking that the array
had no items in it.

The duplicate error is now swallowed by these functions since it is
expected in the case where it is a duplicate badge.