Add Sass, Bourbon, Neat to project
Refactor the Grunt config POJOs

Because the configuration options for the Grunt tasks can get unwieldy,
I've moved the Grunt tasks that were originally there along with the
new Grunt tasks for `grunt-sass` to the new `tools/grunt/config` directory.

Refactor styles to modular files using Sass

This is the first step to preparing the stylesheets for overriding. The
ideas introduced in this patch simply move the styles to separate files
that are imported using `@import` directives in Sass. The cascade has
been respected from the original cascade set in the deleted `application.css`.

Update Sass task to properly integrate into build process

In trying to maintain the build pipeline that was already in place, the Sass task
now does the minimum, *processing Sass/Scss/CSS files*, and allows for minification
from the `cssmin`.