Remove state check from pre-update fn
The logic gate to check for the task state was always returning. This
is because this function is called from the beforeUpdate lifecycle
callback on the Task model. That means, that the changes (including the
state change) have not yet been saved to the database. This results in
every task returning early from this function and therefore no badges
being assigned.

Instead, we’ll rely on the logic gate that is within the lifecycle
callback (the switch/case statement) and assume that anytime this
function is called the task has been appropriately marked as
“completed”.