Add three new methods to Backbone TaskModel;
The `isDraft` and `isSubmission` methods check state on the current task
while `hasBeenSubmitted` checks for the `submittedAt` property to be set
on the current task and if it isn't, then uses the previous two methods
to determine if it has been submitted. Some older tasks before `submittedAt`
was introduced won't have the property set.