Update functionality surrounding editing a task;
This patch adds back the previous functionality of "Saving Changes" to
tasks that do not have a "Draft" state. So any state, i.e. "Submitted",
"Open", "Assigned", etc. These states will allow the user to only "Save
Changes" to a task rather than submit it through submission process. If
a task needs to be "resubmitted" then the task must be put in a "Draft"
state before the ability to "Submit for Review" is restablished.