we shouldn’t depend on task-time-required id
It looks like the code depended on the id of the tag
which was in the database, and that I changed the order
of creation when I re-wrote the scripts that create
the seed data.  Other places in the code it 
used the ‘value’ so I made that change here as well.